fix: removed unwanted type file & added types file respctive compo
