+++ packages/common/src/components/cartIconWithCount/index.ts
@@ -1,0 +1,3 @@
+import CartIconWithCount from './cartIconWithCount'
+
+export default CartIconWithCount
