+++ packages/common/src/components/cutomFilterIcon/index.ts
@@ -1,0 +1,3 @@
+import CustomFilterIcon from './customFilterIcon'
+
+export default CustomFilterIcon
