+++ packages/common/src/components/filter/filter.types.ts
@@ -1,0 +1,5 @@
+export interface FilterPropsModel {
+  handleApplyFilter: (sortBy: string) => void
+  handleResetFilter: () => void
+  handleCancelFilter?: () => void
+}
