+++ packages/common/src/components/filter/index.ts
@@ -1,0 +1,3 @@
+import Filter from './filter'
+
+export default Filter
