+++ packages/common/src/components/homePageContent/index.ts
@@ -1,0 +1,3 @@
+import HomePageContent from './homePageContent'
+
+export default HomePageContent
