+++ packages/common/src/components/index.ts
@@ -1,6 +1,7 @@
-export { default as TopSheet } from './topSheet/topSheet'
-export { default as HomePageContent } from './homePageContent/homePageContent'
-export { default as SearchAndDiscover } from './searchAndDiscover/searchAndDiscover'
-export { default as TopHeader } from './topHeader/topHeader'
-export { default as SubHeader } from './subHeader/subHeader'
-export { default as PaymentMethodSelection } from './paymentMethodSelection/paymentMethodSelection'
+export { default as TopSheet } from './topSheet'
+export { default as HomePageContent } from './homePageContent'
+export { default as SearchAndDiscover } from './searchAndDiscover'
+export { default as TopHeader } from './topHeader'
+export { default as SubHeader } from './subHeader'
+export { default as PaymentMethodSelection } from './paymentMethodSelection'
+export { default as Settings } from './settings'
