+++ packages/common/src/components/paymentDetailsCard/PaymentDetailsCard.module.css
