+++ packages/common/src/components/paymentDetailsCard/index.ts
@@ -1,0 +1,3 @@
+import PaymentDetailsCard from './PaymentDetailsCard'
+
+export default PaymentDetailsCard
