+++ packages/common/src/components/paymentMethodSelection/index.ts
@@ -1,0 +1,3 @@
+import PaymentMethodSelection from './paymentMethodSelection'
+
+export default PaymentMethodSelection
