+++ packages/common/src/components/paymentMethodSelection/paymentMethodSelection.types.ts
@@ -1,0 +1,5 @@
+import { TranslationProps } from '../settings/settings.types'
+
+export interface PaymentMethodSelectionProps extends TranslationProps {
+  handleOrderConfirmation: () => void
+}
