+++ packages/common/src/components/poweredBy/poweredBy.types.ts
@@ -1,0 +1,5 @@
+export interface PoweredByProps {
+  logoSrc: string
+  altText?: string
+  poweredByText: string
+}
