+++ packages/common/src/components/qrCode/QrCode.types.ts
@@ -1,0 +1,3 @@
+export interface QrCodeProps {
+  value: string
+}
