+++ packages/common/src/components/qrCode/index.ts
@@ -1,0 +1,3 @@
+import QrCode from './Qrcode'
+
+export default QrCode
