+++ packages/common/src/components/searchAndDiscover/index.ts
@@ -1,0 +1,3 @@
+import SearchAndDiscover from './searchAndDiscover'
+
+export default SearchAndDiscover
