+++ packages/common/src/components/searchBar/index.ts
@@ -1,0 +1,3 @@
+import searchBar from './searchBar'
+
+export default searchBar
