+++ packages/common/src/components/searchBar/searchBar.types.ts
@@ -1,0 +1,5 @@
+export interface SearchBarProps {
+  searchString: string | string[] | undefined
+  handleChange: Function
+  placeholder?: string
+}
