+++ packages/common/src/components/settings/index.ts
@@ -1,0 +1,3 @@
+import Settings from './settings'
+
+export default Settings
