+++ packages/common/src/components/settings/settings.types.ts
@@ -1,0 +1,4 @@
+export interface TranslationProps {
+  t: (key: string) => string
+  locale?: string
+}
