+++ packages/common/src/components/signIn/index.ts
@@ -1,0 +1,3 @@
+import SignIn from './SignIn'
+
+export default SignIn
