+++ packages/common/src/components/subHeader/index.ts
@@ -1,0 +1,3 @@
+import SubHeader from './subHeader'
+
+export default SubHeader
