+++ packages/common/src/components/topHeader/index.ts
@@ -1,0 +1,3 @@
+import TopHeader from './topHeader'
+
+export default TopHeader
