+++ packages/common/src/components/topHeader/topHeader.types.ts
@@ -1,0 +1,15 @@
+import { TranslationProps } from '../settings/settings.types'
+
+interface HeaderConstants {
+  blackList: {
+    appLogoBlackList: string[]
+    homeIconBlackList: string[]
+    languageIconWhiteList: string[]
+    menuIconWhiteList: string[]
+  }
+}
+
+export interface HeaderProps extends TranslationProps {
+  handleMenuClick?: () => void
+  headerConstants: HeaderConstants
+}
