+++ packages/common/src/components/topSheet/index.ts
@@ -1,0 +1,3 @@
+import TopSheet from './topSheet'
+
+export default TopSheet
