+++ packages/common/src/components/topSheet/topSheet.types.ts
@@ -1,0 +1,8 @@
+import { TranslationProps } from '../settings/settings.types'
+
+export interface TopSheetComponentProps extends TranslationProps {
+  currentAddress: string
+  loadingForCurrentAddress?: boolean
+  currentLocationFetchError?: string
+  searchPlaceholder?: string
+}
